fix: Avoid reallocations for the artifact context buffer
Repeatedly growing and shrinking the buffer has a big potential
of fragmenting the memory space, causing lots of data copying and
even allocation failures. Once we get a big enough buffer for our
operations, we should keep it as long as we need it and then
discard it to free space for other operations.

Ticket: MEN-7562
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>